Separate constant generation to library
