+++ gen_constants_rs.py
@@ -1,5 +1,5 @@
-import os
-import re
+from doxybindgen.constants import generate_constants_in, xml_files_in
+
 import subprocess
 import xml.etree.ElementTree as ET
 
@@ -16,11 +16,7 @@ use std::os::raw::{c_int, c_long};
 use crate::manual::*;
 '''
 
-# MEMO: don't replace `wx` prefix of `wx_GL_COMPAT_PROFILE`
-RE_IDENT = re.compile(r'wx([^_]\w)')
-
 # place wxWidgets doxygen xml files in wxml/ dir and run this.
-generated = set()
 def main():
     outpath = 'wx-base/src/constants.rs'
     with open(outpath, 'w') as f:
@@ -34,243 +30,6 @@ def main():
                 print(line, file=f)
             
     # print(subprocess.check_output(['rustfmt', outpath]))
-
-def generate_constants_in(element):
-    empty = True
-    for define in defines_in(element):
-        empty = False
-        for line in generate_define(define):
-            yield line
-
-    for enum in enums_in(element):
-        empty = False
-        for line in generate_enum(enum):
-            yield line
-    
-    if not empty:
-        yield ''
-
-def xml_files_in(dir):
-    index = os.path.join(dir, 'index.xml')
-    with open(index, 'r') as f:
-        tree = ET.parse(f)
-        root = tree.getroot()
-        for compound in root.findall('./compound'):
-            if has_constants(compound):
-                xml = compound.get('refid') + '.xml'
-                yield os.path.join(dir, xml)
-
-def has_constants(compound):
-    for member in compound.findall("./member"):
-        kind = member.get('kind')
-        if kind in ['define', 'enum']:
-            return True
-    return False
-
-def defines_in(root):
-    memberdefs = root.findall(".//memberdef[@kind='define']")
-    for memberdef in memberdefs:
-        yield memberdef
-
-def enums_in(root):
-    memberdefs = root.findall(".//memberdef[@kind='enum']")
-    for memberdef in memberdefs:
-        yield memberdef
-
-typedefs = [
-    'wxPGVFBFlags',
-]
-blocklist = [
-    # complex defs
-    'Inv_Year',
-    'wxBookCtrl',
-    'wxDISABLE_DEBUG_SUPPORT',
-    'wxDISABLE_ASSERTS_IN_RELEASE_BUILD',
-    'wxEVENT_PROPAGATE_MAX',
-    'wxInvalidDateTime',
-    'wxNullProperty',
-    'wxPG_COLOUR',
-    'wxPG_COLOUR_BLACK',
-    'wxPG_DEFAULT_IMAGE_SIZE',
-    'wxPG_INVALID_VALUE',
-    'wxPG_IT_CHILDREN',
-    'wxPG_LABEL',
-    'wxPG_NULL_BITMAP',
-    'wxPG_PROP_PARENTAL_FLAGS',
-    'wxPGChoicesEmptyData',
-    'wxTLS_TYPE',
-    'wxTreeListEventHandler',
-
-    # wxRichTextRange
-    'wxRICHTEXT_ALL',
-    'wxRICHTEXT_NONE',
-    'wxRICHTEXT_NO_SELECTION',
-
-    # wxSize
-    'wxRICHTEXT_DEFAULT_OVERALL_SIZE',
-    'wxRICHTEXT_DEFAULT_IMAGE_SIZE',
-
-    # wxColour
-    'wxRICHTEXT_DEFAULT_UNFOCUSSED_BACKGROUND',
-    'wxRICHTEXT_DEFAULT_FOCUSSED_BACKGROUND',
-    'wxRICHTEXT_DEFAULT_UNSELECTED_BACKGROUND',
-    'wxRICHTEXT_DEFAULT_TYPE_COLOUR',
-    'wxRICHTEXT_DEFAULT_FOCUS_RECT_COLOUR',
-
-    # min max defs
-    'wxINT8_MIN',
-    'wxINT8_MAX',
-    'wxUINT8_MAX',
-    'wxINT16_MIN',
-    'wxINT16_MAX',
-    'wxUINT16_MAX',
-    'wxINT32_MIN',
-    'wxINT32_MAX',
-    'wxUINT32_MAX',
-    'wxINT64_MIN',
-    'wxINT64_MAX',
-    'wxUINT64_MAX',
-
-    # conflicts with stripping wx-prefix
-    'wxWEST', # use 'LEFT'
-
-    # broken initializer
-    'wxAUI_TBART_OVERFLOW_SIZE',
-    'wxFILE_EXISTS_NO_FOLLOW',
-    'wxPG_PROP_BEING_DELETED',
-]
-def generate_define(e):
-    name = e.findtext('name')
-    if name in generated:
-        return
-    generated.add(name)
-    if name in blocklist or name in typedefs:
-        yield '//  SKIP: %s' % (name,)
-        return
-    initializer = e.find('initializer')
-    if initializer is not None:
-        v = ''.join(initializer.itertext())
-        v = ''.join(map(lambda s: s.lstrip(), v.split('\\\n')))
-        t = 'c_int'
-        if name in long_types:
-            t = 'c_long'
-        if v == 'true' or v == 'false':
-            t = 'bool'
-        elif '.' in v:
-            t = 'f32'
-        elif '"' in v:
-            t = '&str'
-        elif "'" in v:
-            t = 'char'
-        v = re.sub(r'(\d+)[Ll]', r'\1', v)
-        v = re.sub(r'wxString\((".+")\)', r'\1', v)
-        v = re.sub(r'wxS\((".+")\)', r'\1', v)
-        v = re.sub(r'wxT\((".+")\)', r'\1', v)
-        # Don't strip `wx` prefix of string literal (c.f. IMAGE_OPTION_BMP_FORMAT)
-        if '"' not in v:
-            v = RE_IDENT.sub(r'\1', v)
-        name = RE_IDENT.sub(r'\1', name)
-        yield 'pub const %s: %s = %s;' % (name, t, v)
-    else:
-        yield '// NODEF: %s' % (name,)
-
-long_types = [
-    'wxAC_DEFAULT_STYLE',
-    'wxAEDIALOG_STYLE',
-    'wxALWAYS_SHOW_SB',
-    'wxAccObject',
-    'wxBACKINGSTORE',
-    'wxBORDER',
-    'wxBorder',
-    'wxCANCEL',
-    'wxCANCEL_DEFAULT',
-    'wxCAPTION',
-    'wxCHOICEDLG_STYLE',
-    'wxCLIP_CHILDREN',
-    'wxCLIP_SIBLINGS',
-    'wxCLOSE_BOX',
-    'wxCP_DEFAULT_STYLE',
-    'wxDD_DEFAULT_STYLE',
-    'wxDEFAULT_DIALOG_STYLE',
-    'wxDEFAULT_FRAME_STYLE',
-    'wxDOUBLE_BORDER',
-    'wxFULL_REPAINT_ON_RESIZE',
-    'wxGeometryCentre',
-    'wxHLB_DEFAULT_STYLE',
-    'wxHL_ALIGN_CENTRE',
-    'wxHL_CONTEXTMENU',
-    'wxHL_DEFAULT_STYLE',
-    'wxHSCROLL',
-    'wxLB_HSCROLL',
-    'wxMAXIMIZE_BOX',
-    'wxMINIMIZE_BOX',
-    'wxNO_BORDER',
-    'wxOK',
-    'wxPGPropertyFlags',
-    'wxPG_ITERATOR_FLAGS',
-    'wxPG_PROP_ACTIVE_BTN',
-    'wxPG_PROP_MAX',
-    'wxPG_PROP_PASSWORD',
-    'wxPG_PROP_SHOW_FULL_FILENAME',
-    'wxPG_PROP_STATIC_CHOICES',
-    'wxPG_PROP_TRANSLATE_CUSTOM',
-    'wxPG_PROP_USE_CHECKBOX',
-    'wxPG_PROP_USE_DCC',
-    'wxPG_STRING_STORED_FLAGS',
-    'wxPOPUP_WINDOW',
-    'wxRAISED_BORDER',
-    'wxRESIZE_BORDER',
-    'wxRETAINED',
-    'wxSIMPLE_BORDER',
-    'wxSTATIC_BORDER',
-    'wxSTB_DEFAULT_STYLE',
-    'wxSTB_ELLIPSIZE_END',
-    'wxSTB_SHOW_TIPS',
-    'wxSTB_SIZEGRIP',
-    'wxSTC_MASK_FOLDERS',
-    'wxSUNKEN_BORDER',
-    'wxSYSTEM_MENU',
-    'wxTAB_TRAVERSAL',
-    'wxTE_DONTWRAP',
-    'wxTRANSPARENT_WINDOW',
-    'wxTextEntryDialogStyle',
-    'wxVSCROLL',
-    'wxWANTS_CHARS',
-    'wxWINDOW_STYLE_MASK',
-]
-
-def generate_enum(e):
-    name = e.findtext('name')
-    yield '//  ENUM: %s' % (name,)
-    current_initializer = '= 0'
-    count = 0
-    for v in e.findall('enumvalue'):
-        vname = v.findtext('name')
-        if vname in generated:
-            continue
-        generated.add(vname)
-        if vname in blocklist:
-            yield '//  SKIP: %s' % (vname,)
-            continue
-        initializer = v.findtext('initializer')
-        if initializer is None:
-            if count:
-                initializer = '%s + %s' % (current_initializer, count)
-            else:
-                initializer = current_initializer
-            count += 1
-        else:
-            current_initializer = initializer
-            count = 1
-        initializer = initializer.replace('~', '!') # special replacement for wxPATH_NORM_ALL
-        t = 'c_int'
-        if name in long_types:
-            t = 'c_long'
-        if "'" in initializer:
-            t = 'char'
-        vname = RE_IDENT.sub(r'\1', vname)
-        initializer = RE_IDENT.sub(r'\1', initializer)
-        yield 'pub const %s: %s %s;' % (vname, t, initializer)
 
 if __name__ == '__main__':
     main()
